% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{barris}
\alias{barris}
\title{Barcelona neighbourhoods}
\format{
## `barris`
Simple feature collection with 73 features and 2 fields:
\describe{
  \item{osm_id}{openstreetmap item id}
  \item{name}{neighbourhood name}
  ...
}
}
\source{
Openstreetmap Collaborators
}
\usage{
barris
}
\description{
Polygon layer with Barcelona neighbourhoods (barris)
}
\keyword{datasets}
