% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{barris_area_ombra}
\alias{barris_area_ombra}
\title{Number of climate shelters and tree shaded areas by Neighbourhood}
\format{
## `barris_area_ombra`
Simple feature collection with 73 features and 2 field:
\describe{
  \item{area}{total area shaded by trees}
  \item{refugis}{number of climate shelters}
  ...
}
}
\source{
Openstreetmap Collaborators, ICGC & Ajuntament de Barcelona
}
\usage{
barris_area_ombra
}
\description{
Polygon layer with Barcelona neighbourhoods (barris) and its total tree shaded area and climate shelters
}
\keyword{datasets}
