% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{barris_ndvi}
\alias{barris_ndvi}
\title{Avg NDVI by Barcelona neighbourhoods & number of climate shelters}
\format{
## `barris_ndvi`
Simple feature collection with 73 features and 1 field:
\describe{
  \item{ndvi}{normalised difference vegetation index}
  ...
}
}
\source{
Openstreetmap Collaborators & ICGC
}
\usage{
barris_ndvi
}
\description{
Polygon layer with Barcelona neighbourhoods (barris) and its average NDVI
}
\keyword{datasets}
