% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hulls}
\alias{hulls}
\title{Convex hulls of trees less than 10 metres apart and NDVI > .4 over 100 sqm in surface}
\format{
## `hulls`
Simple feature collection with 858 features and 1 field:
\describe{
  \item{grup}{unique hull identifier}
  ...
}
}
\source{
Ajuntament de Barcelona & ICGC
}
\usage{
hulls
}
\description{
Polygon layer with convex hulls of frondous tree clusters
}
\keyword{datasets}
