% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_plot.R
\name{make_map}
\alias{make_map}
\title{make_map}
\usage{
make_map(base = "índex de vegetació total", layers = NA)
}
\arguments{
\item{base}{a base layer}

\item{layer}{layer or layers to be plotted upon base}
}
\value{
a map
}
\description{
makes a plot to be shown in app
}
