% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ndvi}
\alias{ndvi}
\title{Barcelona NDVI}
\format{
## `ndvi`
stars object with 3 dimensions and 1 attribute
}
\source{
<https://geoserveis.icgc.cat/servei/catalunya/ndvi/wms?REQUEST=GetMap&SERVICE=WMS&VERSION=1.3.0&LAYERS=ndvi_color_2022&STYLES=&FORMAT=image/jpeg&BGCOLOR=0xFFFFFF&TRANSPARENT=TRUE&CRS=EPSG:25831&BBOX=420826.2,4574282.1,435505.5,4591026.6&WIDTH=510&HEIGHT=581>
}
\usage{
ndvi
}
\description{
Raster with NDVI in Barcelona
}
\keyword{datasets}
