% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{refugis}
\alias{refugis}
\title{Climate shelters in Barcelona}
\format{
## `refugis`
Simple feature collection with 219 features and 1 field:
\describe{
  \item{name}{shelter name}
  ...
}
}
\source{
<https://opendata-ajuntament.barcelona.cat/data/ca/dataset/xarxa-refugis-climatics>
}
\usage{
refugis
}
\description{
POint layer with all shelters
}
\keyword{datasets}
